+++ examples/nextjs-notifications-custom/src/app/api/liveblocks-auth/route.ts
@@ -1,0 +1,35 @@
+import { Liveblocks } from "@liveblocks/node";
+import { NextRequest } from "next/server";
+import { getUser } from "../../../database";
+import { getExampleUserId } from "../../../example";
+
+// Authenticating your Liveblocks application
+// https://liveblocks.io/docs/authentication
+
+const liveblocks = new Liveblocks({
+  secret: process.env.LIVEBLOCKS_SECRET_KEY as string,
+});
+
+export async function POST(request: NextRequest) {
+  // Get the current user's unique id and info from your database
+  const user = getUser("charlie.layne@example.com");
+
+  if (!user) {
+    return new Response("User not found", { status: 404 });
+  }
+
+  const userId = await getExampleUserId(user.id, request);
+
+  // Create a session for the current user
+  // userInfo is made available in Liveblocks presence hooks, e.g. useOthers
+  const session = liveblocks.prepareSession(userId, {
+    userInfo: user.info,
+  });
+
+  // Use a naming pattern to allow access to rooms with a wildcard
+  session.allow(`liveblocks:examples:*`, session.FULL_ACCESS);
+
+  // Authorize the user and return the result
+  const { body, status } = await session.authorize();
+  return new Response(body, { status });
+}