+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## v2.0.4 (not released yet)
 
+### `@liveblocks/client`
+
+- Add missing type export for `CommentReaction`
+- Don’t attempt to write missing initialStorage keys if the current user has no
+  write access to storage. This will no longer throw, but issue a warning
+  message in the console.
+
 ### `@liveblocks/node`
 
 - Improve `client.initializeStorageDocument()` which can now take a `LiveObject`